adicionado novos recursos

+++ pedido/src/main/java/com/alves/pedido/application/ports/in/categoria/FindCategoriaByIdUseCase.java
@@ -3,5 +3,5 @@ package com.alves.pedido.application.ports.in.categoria;
 import com.alves.pedido.domain.models.Categoria;
 
 public interface FindCategoriaByIdUseCase {
-    Categoria findById(Integer id);
+    Categoria findById(Long id);
 }
