+++ pedido/src/main/java/com/alves/pedido/application/ports/in/cidade/FindCidadeByIdUseCase.java
@@ -3,5 +3,5 @@ package com.alves.pedido.application.ports.in.cidade;
 import com.alves.pedido.domain.models.Cidade;
 
 public interface FindCidadeByIdUseCase {
-    Cidade findById(Integer id);
+    Cidade findById(Long id);
 }
