+++ pedido/src/main/java/com/alves/pedido/application/ports/in/cliente/FindClienteByIdUseCase.java
@@ -3,5 +3,5 @@ package com.alves.pedido.application.ports.in.cliente;
 import com.alves.pedido.domain.models.Cliente;
 
 public interface FindClienteByIdUseCase {
-    Cliente findById(Integer id);
+    Cliente findById(Long id);
 }
