+++ pedido/src/main/java/com/alves/pedido/application/ports/in/pedido/FindPedidoByIdUseCase.java
@@ -3,5 +3,5 @@ package com.alves.pedido.application.ports.in.pedido;
 import com.alves.pedido.domain.models.Pedido;
 
 public interface FindPedidoByIdUseCase {
-    Pedido findById(Integer id);
+    Pedido findById(Long id);
 }
