+++ pedido/src/main/java/com/alves/pedido/application/ports/in/produto/FindProdutoByIdUseCase.java
@@ -3,5 +3,5 @@ package com.alves.pedido.application.ports.in.produto;
 import com.alves.pedido.domain.models.Produto;
 
 public interface FindProdutoByIdUseCase {
-    Produto findById(Integer id);
+    Produto findById(Long id);
 }
