+++ pedido/src/main/java/com/alves/pedido/domain/services/cliente/FindClienteByIdService.java
@@ -12,7 +12,7 @@ public class FindClienteByIdService implements FindClienteByIdUseCase {
     private final FindClienteByIdPort findClienteByIdPort;
 
     @Override
-    public Cliente findById(Integer id) {
+    public Cliente findById(Long id) {
         return findClienteByIdPort.findById(id);
     }
 }
